made soundMode into a u8

+++ src/game/save_file.h
@@ -58,7 +58,7 @@ struct MainMenuSaveData
     // the older the high score is. This is used for tie-breaking when displaying
     // on the high score screen.
     u32 coinScoreAges[NUM_SAVE_FILES];
-    u16 soundMode: 2;
+    u8 soundMode: 2;
 #ifdef WIDE
     u8 wideMode: 1;
 #endif
